Changing references to injectable hCaptcha service and removing previous service
